commit changes

+++ admin/includes/view_all_users.php
@@ -91,12 +91,17 @@
 
 
 
-                        if(isset($_GET['delete'])){
+                        if (isset($_GET['delete'])) {
 
-                            $the_user_id = $_GET['delete'];
+                            if (isset($_SESSION['user_role'])) {
+                                if ($_SESSION['user_role'] == 'admin') {
 
-                            $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-                            $delete_user_query = mysqli_query($connection, $query);
-                            header("Location: users.php");
+                                    $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+
+                                    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+                                    $delete_user_query = mysqli_query($connection, $query);
+                                    header("Location: users.php");
+                                }
+                            }
                         }
                         ?>
